Add permission system

+++ routes/api/jednostka/[jednostkaID]/delete.js
@@ -1,5 +1,8 @@
 import Jednostka from "modules/schemas/jednostka.js";
 
-export default async function({targetJednostka}) {
+export default async function({user, targetJednostka}) {
+	// Check permissions
+	await user.requirePermission(targetJednostka.PERMISSIONS.DELETE)
+	
 	await targetJednostka.deleteOne()
 }
