+++ routes/api/user/[userID]/delete.js
@@ -1,6 +1,9 @@
-import User from "modules/schemas/user.js";
+import User from "modules/schemas/user.js"
+import HTTPError from "modules/server/error.js"
 
-export default async function({targetUser}) {
-	// await User.findByIdAndDelete(userID)
+export default async function({user, targetUser}) {
+	// Check permissions
+	await user.requirePermission(targetUser.PERMISSIONS.DELETE)
+	
 	await targetUser.deleteOne()
 }
